@@ -1,6 +1,8 @@
 "use strict";
-
 module.exports = {
-  PARTIALS: Symbol("partials"),
-  PROFILES: Symbol("profiles")
+    PARTIALS: Symbol("partials"),
+    PROFILES: Symbol("profiles"),
+    OVERRIDE: Symbol("override webpack config partial"),
+    DATA: Symbol("webpack partial data")
 };
+//# sourceMappingURL=constants.js.map
