@@ -1,5 +1,7 @@
+"use strict";
 module.exports = (a, b) => {
-  return (Array.isArray(a) && Array.isArray(b))
-    ? [].concat(b).concat(a)
-    : undefined;
+    return (Array.isArray(a) && Array.isArray(b))
+        ? [].concat(b).concat(a)
+        : undefined;
 };
+//# sourceMappingURL=head-concat-array-merge.js.map
