@@ -1,9 +1,9 @@
-"use strict";
 
-const deleteCustomProps = require("../../lib/delete-custom-props");
-const expect = require("chai").expect;
 
-describe("delete custom props", function () {
+import deleteCustomProps from "../../src/delete-custom-props";
+import { expect } from "chai";
+
+describe("delete custom props", () => {
   it("should preserve __dirname and delete _ prefixed", () => {
     const o = {
       module: {
