@@ -1,9 +1,7 @@
-"use strict";
+import tailConcatArrayMerge from "../../src/tail-concat-array-merge";
+import { expect } from "chai";
 
-const tailConcatArrayMerge = require("../../lib/tail-concat-array-merge");
-const expect = require("chai").expect;
-
-describe("tail concat array", function () {
+describe("tail concat array", () => {
   it("merge two arrays", () => {
     expect(tailConcatArrayMerge([1, 2, 3], [4, 5, 6])).to.deep.equal([1, 2, 3, 4, 5, 6]);
   });
