@@ -251,7 +251,8 @@ export const profiles = {
  */
 export type PlainConfig = Record<string, any>;
 
-import * as WebpackComposer from "webpack-config-composer";
+import WebpackConfigComposer = require("webpack-config-composer");
+const WebpackComposer = WebpackConfigComposer;
 
 /**
  *
